Added Tarsnapper Jobs Configuration Setting
I added a hash-based config option that will enable setting the tarsnapper
jobs.  In order to do this using the built in YAML methods, I had to add
a deep to_hash method to the attribute class, as the default to_hash method
leaves nested Mash objects, which break the serialization.  In addition,
I dropped back the default backup setting to just /etc because of the
way chef builds the options up from deep merges.